+++ src/common_infrastructure/sample_library/sample_library_mdata.cpp
@@ -487,6 +487,17 @@ ReadMdata(Reader& reader, String filepath, ArenaAllocator& result_arena, ArenaAl
     if (auto const o = detail::PostReadBookkeeping(*library, result_arena, scratch_arena); o.HasError())
         PanicIfReached();
 
+    // The PERFORM page didn't used to exist. We need to utilise some vignette on old libraries.
+    if (IsAnyOf(HashFnv1a(library->name),
+                Array {
+                    HashFnv1a("Feedback Loops"),
+                    HashFnv1a("Isolated Signals"),
+                    HashFnv1a("Music Box Suite"),
+                    HashFnv1a("Music Box Suite Free"),
+                    HashFnv1a("Paranormal"),
+                }))
+        library->background_image_vignette_intensity = 100;
+
     // In the MDATA format when velocity-feathering was enabled for an instrument, adjacent velocity layers
     // were automatically made to overlap. We recreate that old behaviour here, taking into account that now
     // velocity feathering is a per-region setting.